Move prismic stuff to prismic service
