+++ events/app/routes.js
@@ -3,7 +3,7 @@ import {renderEvent, renderEventsList} from './controllers';
 
 const r = new Router({ sensitive: true });
 
-r.get('/:preview(preview)?/ev/:id', renderEvent);
-r.get('/ev', renderEventsList);
+r.get('/:preview(preview)?/events/:id', renderEvent);
+r.get('/events', renderEventsList);
 
 export const router = r.middleware();
