+++ prismic-model/curated-lists.json
