+++ prismic-model/people.json
