+++ server/services/prismic.js
@@ -7,11 +7,14 @@ import {
   parseWebcomicDoc,
   asText,
   prismicImage,
-  getPositionInPrismicSeries
+  parseExhibitionsDoc,
+  getPositionInPrismicSeries,
+  parsePromoListItem
 } from './prismic-parsers';
 import {List} from 'immutable';
 import type {PaginatedResults} from '../model/paginated-results';
 import type {ExhibitionPromo} from '../model/exhibition-promo';
+import type {ExhibitionAndRelatedContent} from '../model/exhibition-and-related-content';
 import {PaginationFactory} from '../model/pagination';
 
 type DocumentType = 'articles' | 'webcomics' | 'events' | 'exhibitions';
@@ -266,4 +269,32 @@ async function getEvents(page:number = 1, pageSize:number = 40) {
   ], {orderings: '[my.events.start desc]', page, pageSize});
 
   return eventsList;
+}
+
+export async function getExhibitionAndRelatedContent(id: string, previewReq: ?Request): Promise<?ExhibitionAndRelatedContent> {
+  const exhibition = await getTypeById(previewReq, ['exhibitions'], id, {});
+
+  if (!exhibition) { return null; }
+
+  const ex = parseExhibitionsDoc(exhibition);
+
+  const galleryLevel = exhibition.data.galleryLevel;
+  const promoList = exhibition.data.promoList;
+  const relatedArticles = promoList.filter(x => x.type === 'article').map(parsePromoListItem);
+  const relatedEvents = promoList.filter(x => x.type === 'event').map(parsePromoListItem);
+  const relatedBooks = promoList.filter(x => x.type === 'book').map(parsePromoListItem);
+  const relatedGalleries = promoList.filter(x => x.type === 'gallery').map(parsePromoListItem);
+
+  const sizeInKb = Math.round(exhibition.data.textAndCaptionsDocument.size / 1024);
+  const textAndCaptionsDocument = Object.assign({}, exhibition.data.textAndCaptionsDocument, {sizeInKb});
+
+  return {
+    exhibition: ex,
+    galleryLevel: galleryLevel,
+    textAndCaptionsDocument: textAndCaptionsDocument.url && textAndCaptionsDocument,
+    relatedBooks: relatedBooks,
+    relatedEvents: relatedEvents,
+    relatedGalleries: relatedGalleries,
+    relatedArticles: relatedArticles
+  };
 }
